--- src/main/java/seedu/duke/InvalidCommandException.java
@@ -1,4 +1,0 @@
-package seedu.duke;
-
-public class InvalidCommandException extends Exception {
-}
