+++ src/main/java/seedu/duke/Map.java
@@ -6,6 +6,8 @@ import entity.Location;
 import java.util.Scanner;
 
 import control.loadData;
+import exceptions.InvalidCommandException;
+
 //@@chenling
 public class Map {
     private static loadData dataController = new loadData();
@@ -19,34 +21,41 @@ public class Map {
 
     public static void show_welcome_msg() {
         System.out.println("Welcome to NTU Map \n");
+    }
+
+    public static void promptInput() {
         System.out.println("What would you like to do?");
     }
 
-    public static void show_exit_msg() {
+    public static void showExitMsg() {
         System.out.println("Bye. Hope to see you again soon!\n");
     }
 
-    public static void executeCommand(String input, Command c) throws InvalidCommandException, EmptyInputException {
+    public static void executeCommand(String input, Command c) throws InvalidCommandException {
         switch (c) {
-        case LIST_ALL_LOCATIONS:
-            String location = parser.getLocationsList(input);
-            Feature.listAllLocations(location);
+        case LIST_ALL_LOCATIONS: {
+            String facilityType = parser.getFacilityType(input);
+            Feature.listAllLocations(facilityType);
             break;
-        case SEARCH:
+        }
+        case SEARCH: {
             String facility = parser.getFacilitySearch(input);
             int id = parser.getIdSearch(input);
             searchFacility.query(facility, id);
             break;
-        case SEARCH_IN:
+        }
+        case SEARCH_IN: {
             parser.getBuilding(input);
             break;
-        case FIND_FACILITY:
+        }
+        case FIND_FACILITY: {
             String facilityLocation = parser.getFindFacilityLocation(input);
             Location currentLocation = findNearest.findFacilityByName(dataController, facilityLocation).getLocation();
             String facilityType = parser.getFindFacilityType(input);
             int topK = parser.getTopK(input);
             new findNearest(dataController).findTopKFacility(currentLocation, facilityType, topK);
             break;
+        }
         case INVALID:
             throw new InvalidCommandException();
         default:
@@ -56,10 +65,10 @@ public class Map {
     }
 
     public static void main(String[] args) {
-        assert false : "dummy assertion set to fail";
         show_welcome_msg();
         Command command;
         Scanner in = new Scanner(System.in);
+        promptInput();
         String userInput = ui.getString(in);
 
         while (!parser.isBye(userInput)) {
@@ -80,14 +89,17 @@ public class Map {
             try {
                 executeCommand(userInput, command);
             } catch (InvalidCommandException e) {
-                System.out.println("OOPS!!! I'm sorry, but I don't know what that means :-(");
-            } catch (EmptyInputException e) {
-                System.out.println("OOPS!!! The description of a new task cannot be empty.");
+                System.err.println(e.getMessage());
+//                if (e.getMessage() != null) {
+//                    System.err.println(e.getMessage());
+//                } else {
+//                    System.out.println("OOPS!!! I'm sorry, but I don't know what that means :-(");
+//                }
             }
+            promptInput();
             userInput = ui.getString(in);
-
         }
-
+        showExitMsg();
     }
 
 }
